requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
jinja2
python-dotenv